The Secant Method
This file contains code written in MATLAB, which implements the Secant iteration method to approximate the root of equations. A much more rigorous and robust explanation of the concept in play can be found in the Report file provided for the project. Concepts from undergraduate-level mathematics are involved in some corners of the methodologies practised.