The Regular Falsi Method
This file contains code written in MATLAB, which implements the Regula Falsi iteration method (also The Method of False Position) to approximate the root of equations. A much more rigorous and robust explanation of the concept in play can be found in the Report file provided for the project. Concepts from undergraduate-level mathematics are involved in some corners of the methodologies practised.