Order of Convergence
A code to verify the order of convergence of an iterative sequence (case specific). This code is valid for a very specific sequence given and hence is problem specific. Further, it is only for the purpose of verification. Thus, generic usage will require appropriate tweaks and modifications based on real time requirements.